Windows port now builds, no thanks to a lousy OpenGL header
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@354 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ common/misc.cpp
@@ -21,7 +21,7 @@ using namespace std;
 
 string	get_username() {
 	char username[64];
-	size_t unlen = sizeof(username);
+	DWORD unlen = sizeof(username);
 	if (GetUserName(username,&unlen)) {
 		return string(username);
 	} else {
